- hltypes update for WinRT

+++ trunk/src/rendersystems/DirectX/11/DirectX11_PixelShader.cpp
@@ -45,7 +45,7 @@ namespace april
 			return false;
 		}
 		unsigned char* data = NULL;
-		long size = 0;
+		int32_t size = 0;
 		if (!this->_loadFileData(filename, &data, &size))
 		{
 			hlog::error(april::logTag, "Shader file not found: " + filename);
@@ -69,7 +69,7 @@ namespace april
 			return false;
 		}
 		unsigned char* data = NULL;
-		long size = 0;
+		int32_t size = 0;
 		if (!this->_loadResourceData(filename, &data, &size))
 		{
 			hlog::error(april::logTag, "Shader file not found: " + filename);
